+++ src/main/plugins/discordrpc.js
