+++ src/main/plugins/mpris.js
